Test dataset with 10 mio datapoints counted with cpu approach now works including rendering.

+++ PCViewer/compression/HirarchyCreation.hpp
@@ -70,7 +70,7 @@ namespace compression{
 
     struct IndexCenterFileData{
         float val, min, max;
-        uint32_t offset, size;
+        size_t offset, size;
     };
 
     struct ByteOffsetSize{
